Version 1.1.0
Added support for amending git commits.

Added git prepare-commit-msg and a post-commit hook to handle this. If you are going to upgrade to this version from 1.0.0 you will need to add a column to the timekeeper.tmk file before you start vim, or you will lose the time tracked on OTHER projects that you have been working on. If you are only working on oe project then don't worry. Any future changes will be backwards compatible.